requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
langchain
langchain-groq
groq